servers/asgi-graphql-core/requirements.txt
Click==7.0
graphql-core==3.0.3
h11==0.9.0
httptools==0.1.1
uvicorn==0.11.3
uvloop==0.14.0
websockets==8.1
